refactor: design three-tier architecture, add diagram to docs
