+++ internal/dbrepo/gormrepo/subscription.go
@@ -4,6 +4,8 @@ import (
 	"errors"
 	"time"
 
+	"github.com/vladyslavpavlenko/genesis-api-project/internal/email"
+
 	"github.com/vladyslavpavlenko/genesis-api-project/internal/models"
 
 	"github.com/jackc/pgx/v5/pgconn"
@@ -11,20 +13,23 @@ import (
 
 var ErrDuplicateSubscription = errors.New("subscription already exists")
 
-// SubscriptionRepository is a models.Subscription repository.
-type SubscriptionRepository struct {
+// SubscriberService is a models.Subscription repository.
+type SubscriberService struct {
 	*GormDB
 }
 
 // NewSubscriptionRepository creates a new GormSubscriptionRepository.
-func NewSubscriptionRepository(db *GormDB) *SubscriptionRepository {
-	return &SubscriptionRepository{db}
+func NewSubscriptionRepository(db *GormDB) *SubscriberService {
+	return &SubscriberService{db}
 }
 
 // AddSubscription creates a new Subscription record.
-func (s *SubscriptionRepository) AddSubscription(email string) error {
+func (s *SubscriberService) AddSubscription(emailAddr string) error {
+	if !email.Email(emailAddr).Validate() {
+		return errors.New("invalid email")
+	}
 	subscription := models.Subscription{
-		Email:     email,
+		Email:     emailAddr,
 		CreatedAt: time.Now(),
 	}
 	result := s.DB.Create(&subscription)
@@ -41,7 +46,7 @@ func (s *SubscriptionRepository) AddSubscription(email string) error {
 // GetSubscriptions returns a paginated list of subscriptions. Limit specify the number of records to be retrieved
 // Limit conditions can be canceled by using `Limit(-1)`. Offset specify the number of records to skip before starting
 // to return the records. Offset conditions can be canceled by using `Offset(-1)`.
-func (s *SubscriptionRepository) GetSubscriptions(limit, offset int) ([]models.Subscription, error) {
+func (s *SubscriberService) GetSubscriptions(limit, offset int) ([]models.Subscription, error) {
 	var subscriptions []models.Subscription
 	result := s.DB.Limit(limit).Offset(offset).Find(&subscriptions)
 	if result.Error != nil {
