+++ internal/handlers/repository_test.go
@@ -1,13 +1,58 @@
 package handlers_test
 
 import (
+	"context"
 	"testing"
+
+	"github.com/stretchr/testify/mock"
+	"github.com/vladyslavpavlenko/genesis-api-project/internal/email"
+	"github.com/vladyslavpavlenko/genesis-api-project/internal/models"
 
 	"github.com/vladyslavpavlenko/genesis-api-project/internal/app/config"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/vladyslavpavlenko/genesis-api-project/internal/handlers"
 )
+
+type MockSender struct {
+	mock.Mock
+}
+
+func (m *MockSender) Send(cfg email.Config, params email.Params) error {
+	args := m.Called(cfg, params)
+	return args.Error(0)
+}
+
+type MockSubscriber struct {
+	mock.Mock
+}
+
+func (m *MockSubscriber) GetSubscriptions(_, _ int) ([]models.Subscription, error) {
+	args := m.Called()
+	return args.Get(0).([]models.Subscription), args.Error(1)
+}
+
+func (m *MockSubscriber) AddSubscription(emailAddr string) error {
+	args := m.Called(emailAddr)
+	return args.Error(0)
+}
+
+type MockFetcher struct {
+	mock.Mock
+}
+
+func (m *MockFetcher) Fetch(ctx context.Context, base, target string) (string, error) {
+	args := m.Called(ctx, base, target)
+	return args.String(0), args.Error(1)
+}
+
+func setupServicesWithMocks(subscriber *MockSubscriber, fetcher *MockFetcher, sender *MockSender) *handlers.Services {
+	return &handlers.Services{
+		Subscriber: subscriber,
+		Fetcher:    fetcher,
+		Sender:     sender,
+	}
+}
 
 // TestNewRepo tests the creation of a new repository
 func TestNewRepo(t *testing.T) {
